CRS-2297 Start of refactoring bulk comparison.
